Update go-sdk according to Bison and Cow NEM release It is not final changes. We implemented only change in this branch: https://github.com/nemtech/nem2-sdk-typescript-javascript/tree/release-v0.10.2-beta
* Changed generation of MosaicId and NamespaceId.
* Re-worked MosaicDefinitionTransaction(Schema, serialization, creation).
* Removed code related to previous mosaic implementation.
* Added a new Id of Xem And Xpx (They are same).
* Updated version of transaction and id of SecretProofTransaction.
* Added integration tests which send transaction to blockchain(Integration tests moved to own folder).

+++ sdk/blockchain_model.go
@@ -19,7 +19,7 @@ type BlockInfo struct {
 	NumTransactions       uint64
 	Signature             string
 	Signer                *PublicAccount
-	Version               uint64
+	Version               uint8
 	Type                  uint64
 	Height                *big.Int
 	Timestamp             *big.Int
